+++ docs/CHANGELOG.md
@@ -39,3 +39,4 @@
 - Add ST_MaximumInscribedCircleRadius
 - Add ST_SimplifyPolygonHull
 - Use RelateNG instead of Relate algorithm
+- Add missing implementation on PostGIS datasource wrapper (unwrap and iswrapper)